# docs/swagger.yaml
basePath: /api
definitions:
  dto.AuthResponse:
    properties:
      token:
        description: |-
          JWT authentication token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  dto.BookerRequest:
    properties:
      email:
        description: |-
          Booker's email address
          required: true
          example: john.doe@example.com
        type: string
      full_name:
        description: |-
          Booker's full name
          required: true
          example: John Doe
        type: string
      phone_number:
        description: |-
          Booker's phone number
          required: true
          example: 081234567890
        type: string
    required:
    - email
    - full_name
    - phone_number
    type: object
  dto.BookingResponse:
    properties:
      booking_code:
        description: |-
          Unique booking code
          example: KAI-20240115-001
        type: string
      id:
        description: |-
          Booking unique identifier
          example: 507f1f77bcf86cd799439015
        type: string
      passengers:
        description: List of passengers
        items:
          $ref: '#/definitions/dto.PassengerRequest'
        type: array
      payment_expiry:
        description: |-
          Payment expiry time in YYYY-MM-DD HH:MM format
          example: 2024-01-15 12:00
        type: string
      schedule:
        allOf:
        - $ref: '#/definitions/dto.ScheduleResponse'
        description: Schedule information
      status:
        description: |-
          Booking status
          example: pending
        type: string
      total_price:
        description: |-
          Total price for all passengers
          example: 300000
        type: number
    type: object
...
  dto.CreateBookingRequest:
    properties:
      booker:
        allOf:
        - $ref: '#/definitions/dto.BookerRequest'
        description: |-
          Booker information
          required: true
      passengers:
        description: |-
          List of passengers (minimum 1)
          required: true
        items:
          $ref: '#/definitions/dto.PassengerRequest'
        minItems: 1
        type: array
      schedule_id:
        description: |-
          Schedule ID for the booking
          required: true
          example: 507f1f77bcf86cd799439014
        type: string
    required:
    - booker
    - passengers
    - schedule_id
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        description: |-
          User's email address
          required: true
          example: john.doe@example.com
        type: string
      password:
        description: |-
          User's password
          required: true
          example: password123
        type: string
    required:
    - email
    - password
    type: object
  dto.PassengerRequest:
    properties:
      full_name:
        description: |-
          Passenger's full name
          required: true
          example: Jane Doe
        type: string
      identity_number:
        description: |-
          Passenger's identity number (exactly 16 characters)
          required: true
          example: 1234567890123456
        maxLength: 16
        minLength: 16
        type: string
      seat_code:
        description: |-
          Seat code assigned to the passenger
          required: true
          example: EKO-6 / 9A
        type: string
    required:
    - full_name
    - identity_number
    - seat_code
    type: object
  dto.RegisterRequest:
    properties:
      confirm_password:
        description: |-
          Password confirmation (must match password)
          required: true
          example: password123
        type: string
      email:
        description: |-
          User's email address
          required: true
          example: john.doe@example.com
        type: string
      full_name:
        description: |-
          User's full name (minimum 3 characters)
          required: true
          example: John Doe
        minLength: 3
        type: string
      identity_number:
        description: |-
          User's identity number (exactly 16 characters)
          required: true
          example: 1234567890123456
        maxLength: 16
        minLength: 16
        type: string
      password:
        description: |-
          User's password (minimum 6 characters)
          required: true
          example: password123
        minLength: 6
        type: string
      phone_number:
        description: |-
          User's phone number (minimum 10 characters)
          required: true
          example: 081234567890
        minLength: 10
        type: string
    required:
    - confirm_password
    - email
    - full_name
    - identity_number
    - password
    - phone_number
    type: object
  dto.Response-any:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        description: Response data
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
  dto.Response-array_dto_BookingResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        description: Response data
        items:
          $ref: '#/definitions/dto.BookingResponse'
        type: array
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
...
  dto.Response-dto_AuthResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.AuthResponse'
        description: Response data
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
  dto.Response-dto_BookingResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.BookingResponse'
        description: Response data
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
...
  dto.ScheduleResponse:
    properties:
      arrival_time:
        description: |-
          Arrival time in YYYY-MM-DD HH:MM format
          example: 2024-01-15 16:45
        type: string
      available_seats:
        description: |-
          Number of available seats
          example: 25
        type: integer
      departure_time:
        description: |-
          Departure time in YYYY-MM-DD HH:MM format
          example: 2024-01-15 08:30
        type: string
      destination_station:
        allOf:
        - $ref: '#/definitions/dto.StationResponse'
        description: Destination station information
      duration:
        description: |-
          Journey duration (e.g., "8h 15m")
          example: 8h 15m
        type: string
      id:
        description: |-
          Schedule unique identifier
          example: 507f1f77bcf86cd799439014
        type: string
      origin_station:
        allOf:
        - $ref: '#/definitions/dto.StationResponse'
        description: Origin station information
      price:
        description: |-
          Ticket price
          example: 150000
        type: number
      train:
        allOf:
        - $ref: '#/definitions/dto.TrainInScheduleResponse'
        description: Train information
    type: object
...
  dto.StationResponse:
    properties:
      city:
        description: |-
          City where the station is located
          example: Jakarta
        type: string
      code:
        description: |-
          Station code
          example: GMB
        type: string
      id:
        description: |-
          Station unique identifier
          example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: |-
          Station name
          example: Gambir Station
        type: string
    type: object
  dto.TrainInScheduleResponse:
    properties:
      class:
        description: |-
          Train class
          example: Executive
        type: string
      name:
        description: |-
          Train name
          example: Argo Bromo Anggrek
        type: string
    type: object
...
  dto.UpdateBookingStatusRequest:
    properties:
      status:
        description: |-
          New booking status (e.g., "confirmed", "paid", "cancelled")
          required: true
          example: paid
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: fernandolouis55@gmail.com
    name: Louis Fernando
  description: |-
    This is a comprehensive REST API for Indonesian Railway (KAI) ticket booking system.
    The API provides endpoints for user authentication, station management, train management, schedule search, and booking operations.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: KAI Backend
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.Response-dto_AuthResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized - invalid credentials
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: User login
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with the provided information
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/dto.Response-any'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: Register a new user
      tags:
      - Authentication
  /bookings:
    get:
      consumes:
      - application/json
      description: Retrieve all bookings belonging to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Bookings retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-array_dto_BookingResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Get my bookings
      tags:
      - Bookings
    post:
      consumes:
      - application/json
      description: Create a new train ticket booking for authenticated user
      parameters:
      - description: Booking data
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Booking created successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_BookingResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Create a new booking
      tags:
      - Bookings
  /bookings/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific booking by its ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Booking retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_BookingResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Get booking by ID
      tags:
      - Bookings
  /bookings/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel an existing booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Booking cancelled successfully
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Cancel booking
      tags:
      - Bookings
  /bookings/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a booking (e.g., after payment)
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: New status data
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBookingStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Booking status updated successfully
          schema:
            $ref: '#/definitions/dto.Response-any'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Update booking status
      tags:
      - Bookings
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"